Removed unecessary App and moved NavBar to container
