--- client/src/App.js
@@ -1,9 +1,0 @@
-// import from modules
-import React from 'react';
-
-// import from components
-import AppRouter from './AppRouter';
-
-const App = () => <AppRouter />;
-
-export default App;
